+++ src/main/java/es/us/isa/aml/translators/csp/opl/OPLBuilder.java
@@ -38,7 +38,7 @@ public class OPLBuilder implements IBuilder {
     }
 
     @Override
-    public Float addVersion(Float version) {
+    public Double addVersion(Double version) {
         return version;
     }
